// src/app/interfaces/employee.ts
export interface Employee {
    name: string;
    lastName: string;
    idPerson: string;
    age: number | string;
    email: string;
    phone: string;
    address: string;
    registered: string;
    isEmployee: boolean;
    idEmployee: string;
    job: string;
    hireDate: string;
    isWorking: boolean;
    salary: number | string;
}